Using a menu viewModel instead of repo's and shared pref in fragments

+++ app/src/main/java/com/sigma/sudokuworld/persistence/WordSetRepository.java
@@ -31,7 +31,7 @@ public class WordSetRepository {
         return mAllSets;
     }
 
-    public Set getSet(int setId) {
+    public Set getSet(long setId) {
         return setDao.getSetByID(setId);
     }
